reco/requirements.txt
torch>=2.2.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.7
seaborn>=0.13
pyarrow>=14.0
